requirements.txt
# ============================================================
# Core dependencies for BSS-test framework
# ============================================================
numpy>=1.24
scipy>=1.10
matplotlib>=3.7
pandas>=2.0
pywavelets>=1.4
scikit-learn>=1.3
pyyaml>=6.0
tqdm>=4.65.0
joblib>=1.3.0
